일정 상세 정보 기능 구현

+++ backend/src/main/java/com/ssafy/api/response/ScheduleDetailItem.java
@@ -14,10 +14,10 @@ public class ScheduleDetailItem {
     private String placeUrl;
     private String imageUrl;
     private int reviewCount;
-    private int reviewScore;
+    private Double reviewScore;
     private String tag;
 
-    public ScheduleDetailItem(int scheduleItemId, int day, int jejuPlaceId, String jejuPlaceName, Double latitude, Double longitude, String roadAddress, String placeUrl, String imageUrl, int reviewCount, int reviewScore, String tag) {
+    public ScheduleDetailItem(int scheduleItemId, int day, int jejuPlaceId, String jejuPlaceName, Double latitude, Double longitude, String roadAddress, String placeUrl, String imageUrl, int reviewCount, Double reviewScore, String tag) {
         this.scheduleItemId = scheduleItemId;
         this.day = day;
         this.jejuPlaceId = jejuPlaceId;
